malt_scraper/requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
pyyaml>=6.0.1
fake-useragent>=1.4.0
